Remove xmldoc references to private methods
These have been causing warnings on android CI for a while now that we've just been ignoring. I'd consider looking into why they are only failing there, but to be honest it doesn't really make much sense to reference private methods in the first place like this, so let's just remove them and call it a day.