Use Keycode to map keybord
Replace the switch case with a HashSet that contains InputKey, Key,
SDL_Keycode and SDL_Scancode. A key is assigned to an SDL_Keycode and
if the key is UNKNOWN then it is assigned to an SDL_Scancode as before.
This will allow the shortcuts to be used normally on non-US keyboards.