Fix lack of value transfer happening when decoupled source change involves a non-framed source
Hopefully the comments and tests explain why this is necessary.
Basically, without this the decoupled clock can get in a state where it
is ahead of the new source, but unwilling to seek back into sync due.

One could argue that instead of this, there should be logic osu-side
which ensures the decoupled clock is seeked when the source is changed,
but I don't know.

So yes, at this point I'm basically making up these behaviours for what we
expect from the osu-side. I hope no one else is using these classes, and
also hope I get a chance to rewrite them from scratch in the near
future.