Update log output to show whether a shader was compiled or cached
The previous output was always implying that a compilation happened,
even when it wasn't.

I also removed the pre-compilation log output, as that was mostly there
to gauge performance of shader compiles, which is less of an issue now.
If we still want performance metrics in logs, I'd instead consider
adding a `Stopwatch` and including the elapsed time in the "loaded" log
line.

But this is probably unnecessary.