Refactor `ScrollContainer` scroll delta handling logic to be momre understandable
It was previously not clear the `Y` delta would be used if no `X` delta
was present.

Note that this also adds considerations for precision scroll devices.
Without it, dual-axis scrolls could feel quite weird if the user happens
to scroll a perfectly vertical line (would be unexpectedly converted to
horizontal).