Pool and limit how many GC boxes can be displayed on `FrameStatisticsDisplay` at once
When running on mobile, we've seen gen 0 GCs almost every frame. I'm not
sure if this is good or bad, but ignoring that for a moment, this would
cause opening the frame statistics graphs to slowly grind the game to a
halt as it tries to display this with box markers.

Pooling the boxes reduces the alloc overhead. Custom logic cleans up old
squares to ensure things don't runaway in terms of performance.