Avoid initialising textures when unnecessary
It turns out we were initialising every mipmap level of every textures,
while it is only actually required in scenarios we are not already
replacing every pixel of the texture.

For framework usages, this is for atlases alone.

I've implemented this by changing the `initialisationColour` to be
nullable. For native textures, it has a `null` default, which is perfect
to cover all overrides (`VideoTexture` and `FrameBufferTexture`).

For the `IRenderer` exposed usage, the parameter was already required,
so there is no default. This is good as it ensures a consume isn't going
to see a change in behaviour out-of-the-blue.