Fix Linux/Unix storage paths
`MyDocuments` and `Personal` are
[documented](https://learn.microsoft.com/en-us/dotnet/api/system.environment.specialfolder?view=net-7.0#fields)
as referring to the same thing.

There was a bug with previous versions of .NET where `Personal` would
incorrectly refer to `$HOME` on Unix/Linux. Since 2022-10 (so this change
is likely also required for net7.0), this now correctly refers to
`$HOME/Documents`

The runtime change that caused this is https://github.com/dotnet/runtime/pull/76250.