Refactor/fix GL ubo/ssbo binding
Now looks and behaves much more like the Veldrid implementation.

The extra call to `glUniformBlockBinding` is unlikely to be an issue -
although this is the main drawing function, it's not "hot" per-se (in
that it shouldn't receive more than 1000 draw calls in an extreme
setting).

SSBOs and UBOs don't share the same binding points - `g_GlobalUniforms`
and `g_ColourBuffer` (in the test scene's example) could (and in
practice, do) have an index of 0. What discerns them is the range type,
which is why we need to use `glUniformBlockBinding` for one and
`glShaderStorageBlockBinding` for the other.