Change behaviour of `Storage.Move` to overwrite existing by default
Coming from https://github.com/ppy/osu/actions/runs/5948760882/job/16133189924,
the only way I can see this happening is if the underlying `File.Move`
logic is less atomic than we expect it to be.

In the process of looking into this, I noticed there's an `overwrite`
flag available for `File.Move` which tends to prefer a `rename`
operation internally on unix, rather than `link`.

On windows (the failcase in question) it likely won't help at the end of
the day.

Anyway, personally I think having it overwrite by default is what we
prefer from our API. If that isn't agreed on then closing this without
further consideration is fine.