Disallow overriding `DrawScale`
This was added in a time long forgotten specifically for storyboard
usage. Since then, we have established much better ways of doing this
(easiest is just to make a nested container).

The reasoning for sealing this is that draw scale adds a hidden scale
which is not publicly exposed and can affect all children. It caused me
much frustration in understanding how storyboard elements were a
completely unexpected size.

Given that there was only a few osu-local usages of this, I think we can
remove this as a breaking change.