Include rounded corner radius in android safe area implementation
Android apparently has a separate API for "window cutouts", i.e.
for cameras and the like, and a separate API for "rounded corners",
i.e. devices where the rounded corners of the display will cause some
areas of the theoretically-rectangular display to be obscured.
Because there is no engineering like overengineering, right?
Bonus points for them being usable at different API levels, too.

Disclaimer that this does absolutely nothing on my one test device
that has a rounded display and a high enough API level, but honestly
it may be a device issue since it's not even returning `true`
for `WindowInsets.IsRound`. Maybe it'll help someone, no idea.