Fix multiple instances of a game being allowed to start concurrently
Regressed with https://github.com/ppy/osu-framework/pull/6442.

Was not visible in test coverage because as per the inline comment, this
is an exceptional case where two different processes bind.

I'm not sure how to add test coverage for this, but I have tested
manually on both windows and macOS.