Fix `TextFlowContainer` not laying text out properly on some `TextAnchor` settings
- Closes https://github.com/ppy/osu-framework/issues/5084
- Closes https://github.com/ppy/osu-framework/issues/5499
- Closes https://github.com/ppy/osu-framework/issues/2073
- Closes https://github.com/ppy/osu/issues/8580
- Supersedes / closes https://github.com/ppy/osu-framework/pull/5507

You might ask why I'm bothering to try this now. Well, this came up when
I wanted to use text flow for BSS purposes (and forced me to hack around
it), and came up again in https://github.com/ppy/osu/pull/31970, and I'm
sick of it.

The actual fix is taken verbatim from
https://github.com/ppy/osu-framework/pull/5507, it's just restructured
using the idea of a single nested flow taken from
https://github.com/ppy/osu-framework/pull/5507#discussion_r1023426911.

As such:

Co-authored-by: HiddenNode <101936124+HiddenNode@users.noreply.github.com>