Fix text flow container layout changing after one frame in some cases
Fixes
https://discord.com/channels/90072389919997952/1327149041511043134/1351809773510590494.

Essentially what's happening here is that new lines have weird sizing
rules wherein the height of the new line is relative to the height of
the line preceding it. https://github.com/ppy/osu-framework/pull/6545
broke this because prior to it, the lines responsible for this, namely

	https://github.com/bdach/osu-framework/blob/769a5d5803709a99ef92503d08c192f13e7d913e/osu.Framework/Graphics/Containers/TextFlowContainer.cs#L448-L452

ran in `TextFlowContainer.UpdateAfterChildren()` before the `base` call,
and because `TextFlowContainer` inherited `FillFlowContainer`, changing
the inheritance to composition essentially reordered the call so that
the height set happened *after* `FillFlowContainer`'s layout logic ran,
which is why the layout could be essentially wrong for a frame (with the
new lines having height of zero).

After several hours of trying several hacks to jam this behaviour back
where it should be, I noticed that the entire flow was just broken even
before I started touching it. There were two completely independent
invalidation paths (one in fill flow, one in text flow), and even before
any of my recent  changes, when (ab)using some rarely-used features like
first line indent you could provoke the text flow to start twitching.

So what this does is that it takes the entirety of
`ComputeLayoutPositions()` from fill flow, and adapts it to actually
properly handle all of the features that text flow needs without
post-facto position or margin adjusting hacks.