Fix crash when computing layout of right-anchored autosized text flow
This was failing 10 steps away from the actual problem due to
propagation of weird float values (infinities, NaNs, etc.) The actual
problem was that the offset logic was doing questionable things when the
flow was doing initial autosizing in the X direction. In that case the
maximum width was initialised to `float.MaxValue`, which then in
subsequent calculations got reduced to a very-large-but-finite X
position of everything in the flow, and things just broke.

To circumnavigate that, initialise max size using saner
`float.PositiveInfinity` values (which can be used as markers later),
store the actual widths of the lines rather than the offsets to the
right, and then in the second pass, account for the infinite expansion
due to autosize case by overwriting the max width with the width of the
broadest row.