Fix passthrough input manager mouse position being desynced from parent on first frame
Assume a parenting `InputManager` A and a child
`PassThroughInputManager` B.

On the first frame where B is a child of A, the following sequence of
events occurs:

1. `InputManager` is a `CompositeDrawable`. `CompositeDrawable`s update
   themselves before updating children. As such, A updates itself first.
2. As part of its update logic, A grabs all pending inputs to be
   processed.
3. The collected pending inputs are "applied to current state". The
   process of "applying to current state" is equivalent to passing the
   input to the appropriate input queue in order, until a drawable
   handles the input.
4. While B at this point is a child of A, it is not yet
   `LoadState.Loaded`, because `LoadComplete()` runs as part of the
   first `UpdateSubTree()` call on B, which is yet to occur. Therefore,
   B is omitted from any potentially relevant input queues.
5. Having completed updating itself, A updates its children.
6. As part of that children update, B's `UpdateSubTree()` is called, and
   with it, its `LoadComplete()`.
7. `InputManager.LoadComplete()` has this:

	https://github.com/ppy/osu-framework/blob/576c99b7369432af8d4e6860ac59478e78b5bb48/osu.Framework/Input/InputManager.cs#L184-L186

   which means that not only did it not sync its mouse position with the
   parent even if it has `UseParentInput == true`, it can *actively* use
   a completely different non-zero position due to the aforementioned
   operation.

While all this corrects in the next frame, the entire aforementioned
sequence still means that for the first frame of operation, B will have
a mouse position completely desynced from A, even if it specifies
`UseParentInput == true`.

Because of the aforementioned fact that `CompositeDrawable`s update
themselves before updating children, this allows the *children* to read
the bogus mouse position, too. One such child may be a *replay
recorder*, for instance, thus causing
https://github.com/ppy/osu/issues/28384.

The proposed fix here is a forced synchronisation of mouse position with
the parent on every frame via `Update()`. As, again,
`CompositeDrawable`s update themselves before children, this is
sufficient to avoid exposing the invalid position anywhere lower down.