Fix passthrough input manager so that it works again with game-side usages
You'll probably wisely ask "what is this title? isn't that the tail
wagging the dog?" It is. `PassThroughInputManager` usage in
`RulesetInputManager` sucks. It should not exist.
https://github.com/ppy/osu-framework/pull/6579 broke game-side osu!
touch input mapping (see test failures at
https://github.com/ppy/osu/actions/runs/15438536705/job/43453184186)
because the touch input mapping wants to do its bespoke mouse position
logic on top of `PassThroughInputManager`.

It is my firm belief that `RulesetInputManager` should stop using
`PassThroughInputManager` and instead roll its own bespoke logic. But I
am not going to do that here, because I want to put a framework release
out now rather than in a week, so this attempts to roll out the fix in
https://github.com/ppy/osu-framework/pull/6579 in a slightly different
(read: uglier) way so that it does not break the delicate game-side
usages. If that is not an acceptable outcome, the only other thing I can
offer is a revert.

The last time we had this sort of situation is
`DecoupleableInterpolatingFramedClock` which was also a situation where
the tail wagged the dog for so long and in such inscrutable manner that
eventually it got deleted and rewritten using other classes. Which took
weeks.