Implement runtime checks for Vulkan support
Vulkan is not listed on Android cause it produces
graphical issues. FPS is higher than OpenGL, but
the game is full of blinking artifacts.
Tested on Pixel 7, Pixel 7 Pro and Pixel 8.

On macOS, Vulkan works good through MoltenVK (see
my patches to Veldrid and Vk). Perfomance is on the
same level with Metal, but MoltenVK offers uncapped
to 60 FPS. Also, Vulkan performance is definitely
better than OpenGL.

On Windows and Linux, not all GPUs support Vulkan,
so the check is good to see there imo.