Allow dropdowns to be opened by MouseDown event
Adds `ToggleOnMouseDown` boolean property for Dropdown and DropdownHeader.

The flag will use `MouseDownEvent` instead of `ClickEvent` for toggling the menu.
Dropdown handles `MouseUpEvent`, looking for active hovers and selecting preselected items
if the menu is hovered. Dropdown should close if cursor was released outside the menu and
should stay open if it is released on the `DropdownHeader`.