Fix interpolating clocks' drift recovery being frame rate dependent
This should largely alleviate any remaining concerns of stutters due to
https://github.com/ppy/osu/issues/22488.

In testing, on both windows and macOS the delay for bass to start
playback is around 10 ms. If there is no interpolation, this would cause
a 10 ms stutter.

I've confirmed that interpolation *does* alleviate this already. The
problem is that depending on the update frame rate, it doesn't do a
great job.

With this change, the drift is spread out over multiple frames, which
should make the issue much less visible to an end user (gameplay will
run slightly slower for a brief period rather than outright stopping).