Add primitive to invalidate samples cached by a sample store
The use case for this is in client, in the editor more specifically. I'm
working on better support for custom samples, and a lack of an
invalidation primitive in `SampleStore` meant that even after a sample
was replaced, the old sample would play in the editor until the editor
is exited and re-entered.

A possible client-side alternative of recycling the entire sample store
*is* probably possible, but I think it's excessive.