Free WASAPI if requesting to initialise without it
You might think we can free the device (thus freeing WASAPI) before initialising it,
however this destroys any mixers (and other audio objects) already initialised
under that device, which then start to fail with `Handle` errors.

This is not the same thing as as "re-initialisation", i.e. calling
`InitDevice()` multiple times - we pass in the reinit flag to BASS which
handles this correctly.

Rather than freeing the device, this change tackles the core issue that
WASAPI should be freed if the device is requested to be initialised without it.